Create appolon-framework-1.0.1.css
Appolon is a minimalistic Html5 and CSS3 Framework Template Based for starter
----------
Appolon is also based on 960 grid system. But the core css are about Appolon. We are create this framework to help every starter webdesigner to begin fastly to create webdesign. Appolon is also about webfont icon (Font Awesome Icon). It's minimalist, light, simply and fastest to implement.